Extend the seek bar to the edge of the screen.
Whenever I try to seek when the video is at the start, it ends up
dragging the whole page right, since there's a small gap between the
seek bar and the edge.  This moves the seek bar to the edge to fix this.

The current/total time labels are also changed to left- and right-
aligned, to line up with the seek bar.