Increase the timeout for API requests
I've been running into timeout issues when trying to refresh the song view with a library of > 20k songs on a RasPi 2 and Intel-Atom-based computers. Nothing is wrong with the connection, just takes some time to transfer the data. This just ups the timeout.

By making a contribution to this project, I Mike Megally (@cmsimike) certify with my sign-off line that I agree to the following terms:

Contributor owns, and has sufficient rights to contribute, all source code and related material intended to be compiled or integrated with the source code for the XBMC Remote open source project (the "Contribution") which Contributor has ever delivered, and Giovanni Messina has accepted, for incorporation into the technology made available under the XBMC Remote open source project.

Contributor hereby assigns to Giovanni Messina joint ownership in all worldwide common law and statutory rights associated with the copyrights, copyright application, copyright registration and moral rights in the Contribution to the extent allowable under applicable local laws and copyright conventions. Contributor agrees that this assignment may be submitted by Giovanni Messina to register a copyright in the Contribution. Contributor retains the right to use the Contribution for Contributor's own purposes. This Joint Copyright Assignment supersedes and replaces all prior copyright assignments made by Contributor to Giovanni Messina under the XBMC Remote project.

Contributor is legally entitled to grant the above assignment and agrees not to provide any Contribution that violates any law or breaches any contract.

Signed-off-by: Mike Megally (@cmsimike)