Support pull-to-refresh for TV/Radio recordings
Important remark: The persistence depends on the API method and the parameters. As they are typically same for TV and Radio (the results are filtered later) the object "lifetime" is  commented out for Radio. This gives the required difference which now allows to have separate Radio and Live TV persistence.