Fix unintended section headers in artists list
Section headers were shown even though having <100 results. This was caused by the items not having the key "artist", but only "label". In consequence the default sort method "artist" could not be applied, which activates the section headers. To resolve this a) introduce the key "artist" and b) force refresh when toggling Artists filter mode.