Deselect main menu when entering details from outside of menu structure
When entering item details from outside the menu structure (e.g. long press on an item in a playlist) any active submenu is overwritten with this detail screen. In this case it is best to just deselect the current main menu as users can now enable any desired submenu again with one click.