Start a default after connection was established
Create a new app setting to configure the controller which the app starts with after the connection to Kodi server was established. After connecting search for the desired controller in the main menu and start it. Desired controllers which are not configured as visible in the main menu will not be started. In this case the main menu is shown.