Let MD5 hash for cache also consider App version
This change ensure the cache is refreshed when a new app version is installed. The downside is a re-sync is done for the first time entering a certain menu or view. The benefit is that there is no inconsistency between the cached data format and the expected one. This avoids glitches when presenting the results.