Improved usability of main menu configuration
The setting is now named "Main Menu Items", same as in Kodi UI.
There is no dedicated text anymore (e.g. "Show MUSIC"), but only the item name itself (e.g. "Music"). This simplifies the translation, the consistency with the drop-down list to select the default view and the consistency during not yet finished translations.