Keep playlist visible after snapping back from swipe
Force the NowPlaying view only when entering the view for the first time (viewDidLoad is called). When re-entering the view (viewWillAppear and viewDidAppear) only call handleXBMCPlaylistHasChanged when coming from a child view (e.g. after long-pressing a playlist item and selecting an action) to avoid animating when not needed.
Ensure setting playlistActionView resides in setIphoneInterface and setIpadInterface to avoid overwriting with wrong values.