Fix wrong navbar titles when browsing menus
The mainLabel might be changed when entering a next level while browsing (e.g. Music > Genre > Pop > Genesis). But in case it is not changed the navigation bar title shall show the parameter set's label (e.g. Albums). To achieve this we save the initial mainLabel as rootLabel. If mainLabel changed, show it. If not, show parameters["label"].