Let UI-colorization setting take effect
The UI colorization follows the average color of images (e.g. thumbnails or artwork). Now let the average color be returning a fixed gray color (which supports LightMode and DarkMode) in case the UI-colorization feature is switched off.
As the same method is also used to determine TV logo background a new parameter is introduced to allow ignoring the UI-colorization setting in this case.