Ensure proper fallback to default thumb
The App in few cases receives an invalid path to a thumb. For some item types this invalid path was used to overwrite the default thumb, resulting in no default thumb being loaded at all. Solution is to not overwrite the default thumb to allow the default thumb to fall back to a valid one -- e.g. a folder icon.