Fix flickering playlist when pausing in partymode
The percentage property reports fluctuating values when pausing playback. This causes the playlist to be reloaded while in partymode. As a fix, use the playtime (position in seconds) instead of the percentage. This still allows to detect the start of a new song, which is needed to reload the partymode playlist.