Avoid multiple custom button views on iPad
Introduce a special treatment for custom button view on iPad. This removes an existing topmost custom button view and pushes the latest requested custom button view on top of the stack. This way custom button views will not stack up on multiple presses of the gear icon on iPad remote.