GlobalSearch view must keep choosedTab = 0
GlobalSearch view itself must keep choosedTab = 0 (first tab of this main menu). When selecting or long pressing an item this is overwritten to create the action sheet or child view with the desired content (e.g. music track or movie set). So, after the processing towards a child view or an action sheet is done, chossedTab needs to fall back.