"GetSeasons" does not support parameter "title" for older API versions
The parameter "title" is not support for "GetSeasons" with API < 9.7.0. "GetSeasons" is used to request additional data for section header for the episode view. As the same mapping configuration (mainFields) is used, the easiest approach to resolve this is to change the mapping to use "label" in only this case. With this the problematic "title" parameter can be removed from the JSON request.