Add season/episode info for episode search and episode details
The episode view does not show information on season/episode for each episode as this is given by the section header and the item number. When searching for episodes this information was missing and is now added.
The same information is now also added to the episode details view to reflect information on the season and episode number.