Resolve race condition around delayed JSON API response
JSON API responses can arrive after the NowPlaying view was already removed from the view hierarchy. Adding a check for this resolves a runtime warning when calling cellForRowAtIndexPath on a view which is not in the view hierarchy anymore. In addition, this fixes a problem where storedItemID was set for an old instance of NowPlaying, causing the new instance to not load thumbs and metadata for the current running item anymore.