viewDidDisAppear does not need to end editing or make elements invisible
On iPhone the view is anyway deallocated after reaching viewDidDisappear. On iPad we want to keep the status as this only happens when calling Safari from the App (e.g. following the help link or entering Wikipedia). This fixes an issue where the playlist was empty, the editing got stopped and the NowPlaying overlay was closed on iPad after opening Safari from the App.