Introduce NowPlaying fullscreen for iPad
Moves the play control view out of the toolbar and under the BottomView. This keeps the layout and spacing as it is defined for iPhone.
Changes iPhone's playlist/NowPlaying button to iPad's fullscreen toggle button. On press this sends notification to iPad view controller and then animates to fullscreen view. The method to set the NowPlaying dimension is enhanced to handle both fullscreen and split screen.