Rework and unify power menu for iOS devices
Create a common power action view. Change the approach from using RightViewController for iPhone to an UIAlertView as used on iPad. This allows to introduce a power button in the navigation bar of NowPlaying and the remote screen. As part of the rework this also removes the "LED Torch" in case no server connection is up. Even though a loss of functionality, this is in line with iPad. All now unused configurations in AppDelegate are removed.