Rework searchbar logic
Only create self.searchController once in viewDidLoad.
self.searchController is the UISearchController which is used for an active search and always displayed detached (floating on top of list). To activate a search a dedicated UISearchController + transparent overlayed UIView is created for list view and collection view each. The overlay has a UITapGestureRecognizer which triggers a method to show and activate the central self.searchController. This allows keep the same look of the searchbar.
Only create the inactive searchbar + overlay once for list view and once for collection view in viewDidLoad.
Some refactoring to be able to unify the look of the inactive searchbar (the one with the overlay) and the active self.searchController