iPad menu layout improvement
Menu and playlist now share same vertical gray separator of width 0.5. In consequence playlistLeftShadow has been removed. A define has been introduced to set the separator width.
Also remove the horizontal separator which was added via a semi-transparent thin UIView in the xib. This is now in line with the iPhone layout of the main menu.