Re-introduce a UIView into DetailViewController
This UIView is required to be able to set clipToBounds to hide the inactive elements (e.g. more items menu or grid view) out of sight on iPad. Directly setting a mask for the self.view is not possible as this would not allow to add visible shadows (see addShadowsToView) around the masked area.