Let JSON commands be dropped if server is not connected
This avoids commands are queueing up if the server is temporarily disconnected. In case of not being connected an error is reported back.
JSON commands not related to heartbeat do not need and shall not use a specific timeout. This allows to easily differentiate the heartbeat JSON commands, which need to be executed in any case, from the ones which can be dropped.