Fix EXIF details font toggling between bold and normal
A bold font is used to display EXIF data, using an attributed text inside the label itemDescription. Therefore it is causing trouble when resetting the font and font size of itemDescription when the NowPlaying dimensions were change, e.g. when switching to fullscreen or back. Instead we only keep the desired font size and set the font inside the conditions for VIDEO and MUSIC playback.