Layout fixes for channel list after viewing recordings
Since the main menu supports separated Live TV and Radio the same instance of DetailVC is used for channel lists and recordings. For recording views a different height and font size for the genre label is used as well as hiding the runtime. Due to cell reuse the genre label height, font size and runtime label's visibility must be explicitly set for channel list.