Fix non-working grid view for settings
Ensure collectionView is always initialized in viewDidLoad, also if the first tab does not support grid view. Otherwise collection view is not initialized for other tabs of the same controller instance. Also ensure the change of the list view's start frame is only altered after the collection view was initialized. Both are relevant for the settings controller view.