Ignore autoscroll after user dragged playlist, but with timeout
When user dragged the playlist to view a different section of the playlist, the playlist should not automatically scroll the current playing item into middle of the view. Instead, the autoscroll is ignored until a timeout kicks in to fall back to autoscroll which again places the current playing item into the middle of the view.