Major rework of iPad StackScrollViewController
The main method to move and place the views (handlePanFrom) has been rewritten and is now a lot less complex and easier to maintain. The use of bounce has been rewritten to use the right direction and apply it to all involved views. In addition many non needed methods and ivars have been removed as well.
As a consequence also hitTest (ViewControllerIPad) needs an update as it hard codes the index inside StackScrollViewController.