Fast server connection for iPad
Last Kodi server is already loaded in AppDelegate before ViewController IPad is initialized. To immediately show the HostManagementVC in case of no last server is defined, or the server cannot be connected, starting the rcpJSONRPC can only be done in viewDidAppear. Otherwise the popover cannot be shown and the popover is delayed by several seconds.