Smoother transition to a configured "start view"
To avoid sliding out the HostManagementVC and sliding in the targeted "start view", the HostManagementVC is hidden, if "start view" is configured different to main menu. To ensure the HostManagementVC is shown in case of connection problems it now registers to "TcpJSONShowSetup" which requests to show / unhide the HostManagementVC in such case.