Reimplement starting playback in ShowInfoVC
Avoid to use Playlist.Clear, Playlist.Add and starting to play a playlist index. Instead, directly call Player.Open with the item, which then allows to resume the position via API. As Player.Open anyway will clear and update the playlist, this is a much simpler and more robust way to start or resume the playback.