Reimplement resume playback
Avoid to use Playlist.Clear, Playlist.Add and Player.Open with playlist index. Instead, directly call Player.Open with the item, which then allows to resume the position via API. As Player.Open anyway cause the playlist to be cleared and updated, this is a much simpler and more robust way to resume the playback.