Rework tab handling to avoid race conditions
choosedTab represents the selected view in a menu, e.g. Albums, Artists or Genre. Instead of changing the ivar choosedTab when processing actions in Global Search, it is better to use the local variable activeTab. This avoids race conditions and allows to remove special handling when coming back to Global Search from a child view or after processing an action in Global Search. This requires the implementation of getActiveTab which returns the targeted activeTab for a given item.