Update progress/recording indicator for Live TV and Radio
Update the indicators for progress and recording to a progress bar. The progress bar is located on the left side, uses dark gray background and Kodi blue for the progress. Total runtime is located right aligned and under the bar. The recording indicator is located in a reserved area left of the runtime.
In the EPG the highlighting is done by changing the cell background and letting the start time label use a bigger font size and label color.
Introduces the class ProgressBarView for future use by other views as well. This is used as subview by the new class BroadcastProgressView which implements the layout described above and replaces the xib-defined progress view.