Reintroduce smooth animation of progressbar
The animation is only shown when the progressbar is updated from width of 0 (= start value) to any non-zero value. Effectively, the animation is only done the first time a playlist is loaded during an already advanced playback.
The xib starts with progress = 11 to visualize the progress bar's function in the xib.