Rework selection strategy
Remove selection when showing an action sheet. The details for the selected item are anyway shown on top of the action sheet, and on iPad there is a pointer to the touched origin.
At the same time select an item on tap and keep it selected when entering a next level view. This visually emphasizes the selected item when showing two view side-by-side as it is done on iPad.