Rework initialization of main iPhone ViewControllers
InitialSlidingVC is the main controller for iPhone. It has the top, the underleft and underright controller. When starting, the top controller is either the default HostManagementVC or any other menu controller defined by the user settings. The underleft controller is always the main menu (= MasterVC). These two controllers (top and underleft) should be created and assigned when InitialSlidingVC is initialized.