Do not call setNeedsLayout in DetailVCs configureView
This was recently introduced when fixing an issue regaring wrong layout of NowPlaying when entering via DetailVC.
But setNeedsLayout is only needed in NowPlaying and not in DetailVC. In DetailVC this is even has an unwanted side effect (self.navigationController is nil in viewDidLoad). Hence we remove this call again.