Rework and fix using the onoff switch
Do not use the tag anymore to identify which custom button was selected. Instead, use convertPoint and indexPathForRowAtPoint to read the row. This simplifies the logic to move a custom button's position. It also fixes custom buttons with onoff switches to fail after another custom button was moved on above of them.