Revise load and update strategy for custom buttons
Custom buttons are common for all servers, but the status of toggle buttons might be different. Therefore there are two different actions required: load and update. Load is about loading the custom buttons which the user created. This is needed when the view is loaded (viewDidLoad), when the custom buttons were added (UIInterfaceCustomButtonAdded) or when custom buttons were deleted (deleteCustomButton). Update is not changing the custom buttons, but updates their state, like the toggle button state. This needs to be done via reloadData on connectionSuccess, when a new server got selected (XBMCServerHasChanged) and each time the buttons were freshly loaded.