Improve showing gradient and labels while loading images
Ensure default gradient and text colors are shown during image loading. This requires graidentLayer to be created and added to view in layoutSectionView instead of setViewColor. setViewColor is called twice: first time to show the default colors while loading, second time to update the colors based on the loaded image.