Add special processing for streams
Assumption is that a stream will have a valid host in NSURLComponents. In this case percent encoding is removed and "+" is replaced by " " to take into account that some plugins, like "ORF Radiothek", share query-like label via the API.
The host name is shown as detailing text for both playlist and NowPlaying screen.